app/services/research_service.py
from typing import List, Dict, Any, Optional
import logging
from langchain.vectorstores import Chroma
from langchain.embeddings import OpenAIEmbeddings
from langchain.text_splitter import RecursiveCharacterTextSplitter
from langchain.document_loaders import TextLoader, PDFLoader, WebBaseLoader
from langchain.chains import RetrievalQA
from langchain.chat_models import ChatOpenAI
import chromadb
import os
from datetime import datetime
import json
import aiohttp
import asyncio
from bs4 import BeautifulSoup

class ResearchService:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.embeddings = OpenAIEmbeddings()
        self.db = Chroma(
            persist_directory="./knowledge_base",
            embedding_function=self.embeddings
        )
        self.llm = ChatOpenAI(temperature=0.1)
        self.research_cache = {}

    async def add_research_document(self, content: str, metadata: Dict[str, Any]) -> bool:
        """
        Add a research document to the knowledge base
        """
        try:
            text_splitter = RecursiveCharacterTextSplitter(
                chunk_size=1000,
                chunk_overlap=200
            )
            texts = text_splitter.split_text(content)
            
            # Add timestamp and source
            metadata["timestamp"] = datetime.now().isoformat()
            
            self.db.add_texts(
                texts=texts,
                metadatas=[metadata] * len(texts)
            )
            return True
        except Exception as e:
            self.logger.error(f"Error adding research document: {str(e)}")
            return False

    async def query_knowledge_base(self, query: str, filters: Optional[Dict] = None) -> Dict[str, Any]:
        """
        Query the knowledge base for relevant information
        """
        try:
            retriever = self.db.as_retriever(
                search_kwargs={"filter": filters} if filters else {}
            )
            
            qa_chain = RetrievalQA.from_chain_type(
                llm=self.llm,
                chain_type="stuff",
                retriever=retriever
            )
            
            response = qa_chain({"query": query})
            return {
                "status": "success",
                "answer": response["result"],
                "sources": response.get("source_documents", [])
            }
        except Exception as e:
            self.logger.error(f"Error querying knowledge base: {str(e)}")
            return {"status": "error", "message": str(e)}

    async def generate_research_report(self, topic: str, depth: str = "comprehensive") -> Dict[str, Any]:
        """
        Generate a research report on a specific topic
        """
        try:
            # Query knowledge base
            base_knowledge = await self.query_knowledge_base(
                f"Provide a {depth} analysis of {topic}"
            )
            
            # Generate report structure
            report_structure = {
                "title": f"Research Report: {topic}",
                "timestamp": datetime.now().isoformat(),
                "executive_summary": "",
                "methodology": "",
                "findings": "",
                "conclusions": "",
                "recommendations": "",
                "references": []
            }
            
            # Generate each section using the LLM
            prompts = {
                "executive_summary": f"Write an executive summary for a research report on {topic}",
                "methodology": "Describe the research methodology used",
                "findings": f"Provide detailed findings about {topic}",
                "conclusions": "Draw conclusions from the findings",
                "recommendations": "Make recommendations based on the conclusions"
            }
            
            for section, prompt in prompts.items():
                response = await self.llm.agenerate([prompt])
                report_structure[section] = response.generations[0][0].text
            
            return {
                "status": "success",
                "report": report_structure
            }
        except Exception as e:
            self.logger.error(f"Error generating research report: {str(e)}")
            return {"status": "error", "message": str(e)}

    async def monitor_research_trends(self, keywords: List[str]) -> Dict[str, Any]:
        """
        Monitor and analyze research trends related to keywords
        """
        try:
            trends = {}
            for keyword in keywords:
                # Query academic sources (placeholder for actual API calls)
                sources = [
                    "arxiv",
                    "google_scholar",
                    "research_gate"
                ]
                
                keyword_data = {
                    "recent_papers": [],
                    "trending_topics": [],
                    "citation_metrics": {},
                    "related_concepts": []
                }
                
                # Store in cache with timestamp
                self.research_cache[keyword] = {
                    "data": keyword_data,
                    "timestamp": datetime.now().isoformat()
                }
                
                trends[keyword] = keyword_data
            
            return {
                "status": "success",
                "trends": trends
            }
        except Exception as e:
            self.logger.error(f"Error monitoring research trends: {str(e)}")
            return {"status": "error", "message": str(e)}

    async def generate_citation(self, content: Dict[str, Any], style: str = "APA") -> str:
        """
        Generate academic citations in various formats
        """
        try:
            citation_prompt = f"Generate a {style} citation for the following content: {json.dumps(content)}"
            response = await self.llm.agenerate([citation_prompt])
            return response.generations[0][0].text
        except Exception as e:
            self.logger.error(f"Error generating citation: {str(e)}")
            return ""
