app/services/technical_service.py
from typing import Dict, Any, List, Optional
import logging
import boto3
from azure.storage.blob import BlobServiceClient
from google.cloud import storage
import json
import asyncio
from datetime import datetime
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.cluster import KMeans
from transformers import pipeline

class TechnicalService:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.cloud_clients = {}
        self.nlp_pipeline = pipeline("text2text-generation")
        
    async def init_cloud_clients(self, credentials: Dict[str, Any]):
        """
        Initialize cloud service clients
        """
        try:
            if "aws" in credentials:
                self.cloud_clients["aws"] = boto3.client(
                    "s3",
                    aws_access_key_id=credentials["aws"]["access_key"],
                    aws_secret_access_key=credentials["aws"]["secret_key"]
                )
            
            if "azure" in credentials:
                self.cloud_clients["azure"] = BlobServiceClient.from_connection_string(
                    credentials["azure"]["connection_string"]
                )
            
            if "gcp" in credentials:
                self.cloud_clients["gcp"] = storage.Client.from_service_account_info(
                    credentials["gcp"]
                )
            
            return True
        except Exception as e:
            self.logger.error(f"Error initializing cloud clients: {str(e)}")
            return False

    async def analyze_cloud_costs(self, provider: str, timeframe: str) -> Dict[str, Any]:
        """
        Analyze and optimize cloud infrastructure costs
        """
        try:
            if provider not in self.cloud_clients:
                raise ValueError(f"Cloud provider {provider} not initialized")
            
            # Placeholder for actual cost analysis logic
            analysis = {
                "total_cost": 0,
                "cost_by_service": {},
                "optimization_recommendations": [],
                "potential_savings": 0
            }
            
            return {
                "status": "success",
                "analysis": analysis
            }
        except Exception as e:
            self.logger.error(f"Error analyzing cloud costs: {str(e)}")
            return {"status": "error", "message": str(e)}

    async def generate_infrastructure_code(self, 
                                        requirements: Dict[str, Any]) -> Dict[str, Any]:
        """
        Generate Infrastructure as Code (IaC) based on requirements
        """
        try:
            # Template selection based on requirements
            template_type = requirements.get("type", "kubernetes")
            cloud_provider = requirements.get("provider", "aws")
            
            # Placeholder for actual code generation logic
            generated_code = {
                "main.tf": "# Terraform configuration\n",
                "variables.tf": "# Variables\n",
                "outputs.tf": "# Outputs\n"
            }
            
            return {
                "status": "success",
                "code": generated_code,
                "documentation": "# Infrastructure Documentation\n"
            }
        except Exception as e:
            self.logger.error(f"Error generating infrastructure code: {str(e)}")
            return {"status": "error", "message": str(e)}

    async def analyze_system_performance(self, 
                                      metrics: List[Dict[str, Any]]) -> Dict[str, Any]:
        """
        Analyze system performance metrics and provide recommendations
        """
        try:
            df = pd.DataFrame(metrics)
            
            # Basic statistical analysis
            analysis = {
                "summary_stats": df.describe().to_dict(),
                "correlations": df.corr().to_dict(),
                "anomalies": []
            }
            
            # Anomaly detection using clustering
            scaler = StandardScaler()
            scaled_data = scaler.fit_transform(df.select_dtypes(include=[np.number]))
            kmeans = KMeans(n_clusters=3)
            clusters = kmeans.fit_predict(scaled_data)
            
            # Generate recommendations
            recommendations = [
                "Optimize resource allocation",
                "Scale infrastructure based on usage patterns",
                "Implement caching for frequently accessed data"
            ]
            
            return {
                "status": "success",
                "analysis": analysis,
                "recommendations": recommendations
            }
        except Exception as e:
            self.logger.error(f"Error analyzing system performance: {str(e)}")
            return {"status": "error", "message": str(e)}

    async def generate_api_documentation(self, 
                                      code_base: Dict[str, Any]) -> Dict[str, Any]:
        """
        Generate API documentation from codebase
        """
        try:
            # Extract API endpoints and schemas
            documentation = {
                "openapi": "3.0.0",
                "info": {
                    "title": code_base.get("title", "API Documentation"),
                    "version": code_base.get("version", "1.0.0")
                },
                "paths": {},
                "components": {
                    "schemas": {}
                }
            }
            
            # Generate documentation for each endpoint
            for endpoint in code_base.get("endpoints", []):
                path_doc = self.nlp_pipeline(
                    f"Document API endpoint: {json.dumps(endpoint)}"
                )[0]["generated_text"]
                
                documentation["paths"][endpoint["path"]] = json.loads(path_doc)
            
            return {
                "status": "success",
                "documentation": documentation
            }
        except Exception as e:
            self.logger.error(f"Error generating API documentation: {str(e)}")
            return {"status": "error", "message": str(e)}

    async def security_audit(self, 
                           system_config: Dict[str, Any]) -> Dict[str, Any]:
        """
        Perform security audit and compliance check
        """
        try:
            audit_results = {
                "timestamp": datetime.now().isoformat(),
                "compliance": {
                    "gdpr": [],
                    "hipaa": [],
                    "iso27001": []
                },
                "vulnerabilities": [],
                "recommendations": []
            }
            
            # Analyze configurations
            for component, config in system_config.items():
                # Placeholder for actual security analysis logic
                audit_results["vulnerabilities"].extend([
                    {
                        "component": component,
                        "severity": "medium",
                        "description": "Default configuration used"
                    }
                ])
            
            return {
                "status": "success",
                "audit": audit_results
            }
        except Exception as e:
            self.logger.error(f"Error performing security audit: {str(e)}")
            return {"status": "error", "message": str(e)}
