requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
schedule==1.2.1
openai==1.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dateutil==2.8.2
pyautogui==0.9.54
keyboard==0.13.5
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
python-jose[cryptography]==3.3.0
emails==0.6
jinja2==3.1.2
selenium==4.15.2
playwright==1.40.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
langchain==0.0.350
chromadb==0.4.18
beautifulsoup4==4.12.2
pymongo==4.6.1
redis==5.0.1
boto3==1.33.6
azure-storage-blob==12.19.0
google-cloud-storage==2.13.0
python-gitlab==3.15.0
slack-sdk==3.26.1
azure-identity==1.15.0
